# docs/swagger.yaml
##############################################
# this is an example of the link curation API
# as a demonstration of an API spec in YAML
##############################################

swagger: '2.0'
info:
  title: Link Curation API
  description: We support three basic APIs- Questions, Answers and User Account Services
  version: "1.0.0"

# the domain of the service
host: localhost:5000

# array of all schemes that your API supports
schemes:
  - http

# will be prefixed to all paths
basePath: /v1
produces:
  - application/json

# All different endpoints
paths:

  ##############################################
  # User
  ##############################################
  
  /user:
  
    ############################################## POST
    
    post:
      summary: Register a new user
      description: |
        Create new user account for a curator
        
        Examples:
        - curl -X POST http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"username":"nilayvac@usc.edu","password":"linkCuration"}'
        - curl -X POST http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"username":"ksureka@usc.edu","password":"linkCuration"}'
      parameters:
        - name: payload
          in: body
          description: Input Payload
          required: true
          schema:
            $ref: '#/definitions/UserPostInputPayload'
      tags:
        - User
      responses:
        200:
          description: Profile information for a user
          schema:
            type: array
            items:
              $ref: '#/definitions/UserPostOutputPayload'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    
    ############################################## GET
    
    get:
      summary: Login existing user and get time bound API token
      description: |
        Authenticates user credentials and returns time sensitive API access token. This token can be used instead of credentials to access protected resources
        
        Examples:
        - curl -X GET http://localhost:5000/v1/user -u nilayvac@usc.edu:linkCuration
        - curl -X GET http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"duration":3600}' -u nilayvac@usc.edu:linkCuration
        - curl -X GET http://localhost:5000/v1/user -u ksureka@usc.edu:linkCuration
        - curl -X GET http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"duration":3600}' -u ksureka@usc.edu:linkCuration
      parameters:
        - name: payload
          in: body
          description: Input Payload
          required: true
          schema:
            $ref: '#/definitions/UserGetInputPayload'
      tags:
        - User
      responses:
        200:
          description: Profile information of user along with token details
          schema:
            type: array
            items:
              $ref: '#/definitions/UserGetOutputPayload'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    
    ############################################## PUT
    
    put:
      summary: Update user profile 
      description: |
        Update user profile details like name of the user, tags and rating. User Authentication required to access this API. At least one of the three paraneters is required
        
        Exampls:
        - curl -X PUT http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"name":"Nilay Chheda","rating":5,"tags":["saam","ulan"]}' -u nilayvac@usc.edu:linkCuration
        - curl -X PUT http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"name":"Nilay Chheda","rating":5,"tags":["saam","ulan"]}' -u eyJhbGciOiJIUzI1NiIsImV4cCI6MTQ2MDU5OTUxNCwiaWF0IjoxNDYwNTk1OTE0fQ.eyJpZCI6Mn0.FSoeJkqaV1Zlc1XjDu5fcI3fmRSHD1OMhm-M8sKOHE8:x
        - curl -X PUT http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"name":"Karishma Sureka","rating":5,"tags":["dbpedia","npg"]}' -u ksureka@usc.edu:linkCuration
        - curl -X PUT http://localhost:5000/v1/user -H "Content-Type: application/json" -d '{"name":"Karishma Sureka","rating":5,"tags":["dbpedia","npg"]}' -u eyJhbGciOiJIUzI1NiIsImV4cCI6MTQ2MDU5OTUxNCwiaWF0IjoxNDYwNTk1OTE0fQ.eyJpZCI6Mn0.FSoeJkqaV1Zlc1XjDu5fcI3fmRSHD1OMhm-M8sKOHE8:x
        
      parameters:
        - name: payload
          in: body
          description: Input Payload
          required: true
          schema:
            $ref: '#/definitions/UserPutInputPayload'
      tags:
        - User
      responses:
        200:
          description: Profile information of user along with token details
          schema:
            type: array
            items:
              $ref: '#/definitions/UserPutOutputPayload'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
        
  ##############################################
  # Question
  ############################################## 
  
  /question:
  
    ############################################## GET
  
    get:
      summary: Serve relevant questions from database
      description: |
        This Question endpoint returns set of questions to client based on different paraneters associated 
        with the user and user's previous history. User Authentication required to access this API
        
        Examples: 
        - curl -X GET http://localhost:5000/v1/question -u nilayvac@usc.edu:linkCuration
        - curl -X GET http://localhost:5000/v1/question -H "Content-Type: application/json" -d '{"stats":"True"}' -u nilayvac@usc.edu:linkCuration
        - curl -X GET http://localhost:5000/v1/question -H "Content-Type: application/json" -d '{"bulk":10}' -u nilayvac@usc.edu:linkCuration
        - curl -X GET http://localhost:5000/v1/question -H "Content-Type: application/json" -d '{"bulk":10,"stats":"True"}' -u nilayvac@usc.edu:linkCuration
        - curl -X GET http://localhost:5000/v1/question -H "Content-Type: application/json" -d '{"bulk":10,"stats":"True"}' -u eyJhbGciOiJIUzI1NiIsImV4cCI6MTQ2MDU5OTUxNCwiaWF0IjoxNDYwNTk1OTE0fQ.eyJpZCI6Mn0.FSoeJkqaV1Zlc1XjDu5fcI3fmRSHD1OMhm-M8sKOHE8:x
        
      parameters:
        - name: payload
          in: body
          description: Input Payload
          required: true
          schema:
            $ref: '#/definitions/QuestionGetInputPayload'
      tags:
        - Question
      responses:
        200:
          description: An array of questions
          schema:
            type: array
            items:
              $ref: '#/definitions/QuestionGetOutputPayload'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
    ############################################## POST
    post:
      summary: Create new question(s) (aka pair(s)) in the database
      description: |
        This Question endpoint allows external software entity like dedupe to create new questions. No authentication required. 
        
        Examples:
        - curl -X POST http://localhost:5000/v1/question -H "Content-Type: application/json" -d '{"uri1":"http://vocab.getty.edu/ulan/50001234","uri2":"http://edan.si.edu/saam/id/person-institution/1234","dedupe":{"ver":"1.0","revision":"4","score":"0.45"}}'
        - curl -X POST http://localhost:5000/v1/question -H "Content-Type: application/json" -d '{"bulk":count n,"payload":[{},{},... n entries]}'
        - curl -X POST http://localhost:5000/v1/question -H "Content-Type: application/json" -d '{"bulk":2,"payload":[{"uri1":"http://vocab.getty.edu/ulan/50002345","uri2":"http://edan.si.edu/saam/id/person-institution/2345","dedupe":{"ver":"1.0","revision":"4","score":"0.45"}},{"uri1":"http://vocab.getty.edu/ulan/50003456","uri2":"http://edan.si.edu/saam/id/person-institution/3456","dedupe":{"ver":"1.0","revision":"4","score":"0.85"}}]}'
        
      parameters:
        - name: payload
          in: body
          description: Input Payload
          required: true
          schema:
            $ref: '#/definitions/QuestionPostInputPayload'
        
      tags:
        - Question
      responses:
        200:
          description: Curation interface data about Question(s)
          schema:
            type: array
            items:
              $ref: '#/definitions/QuestionPostOutputPayload'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  ##############################################
  # Answer
  ##############################################
  
  /answer:
  
    ############################################## PUT
    
    put:
      summary: Submit an answer
      description: |
        Submit human curated answer related to a question. User Authentication required to access this API.
        
        Examples:
        - curl -X PUT http://localhost:5000/v1/answer -H "Content-Type: application/json" -d '{"value":"1","comment":"Just another comment","qid":"570eec1df6bf2d1e58a88479"}' -u nilayvac@usc.edu:linkCuration
        - curl -X PUT http://localhost:5000/v1/answer -H "Content-Type: application/json" -d '{"value":"1","comment":"Just another comment","qid":"570eec1df6bf2d1e58a88479"}' -u ksureka@usc.edu:linkCuration
        - curl -X PUT http://localhost:5000/v1/answer -H "Content-Type: application/json" -d '{"value":"1","comment":"Just another comment","qid":"570eeba7f6bf2d1e58a88471"}' -u eyJhbGciOiJIUzI1NiIsImV4cCI6MTQ2MDU5OTQ4MSwiaWF0IjoxNDYwNTk1ODgxfQ.eyJpZCI6MX0.PyLZSP7q0iWh-n1yazz8p23kO3cUJtmQbGucjvTYtOg:x
        - curl -X PUT http://localhost:5000/v1/answer -H "Content-Type: application/json" -d '{"value":"1","comment":"Just another comment","qid":"570eeba7f6bf2d1e58a88471"}' -u eyJhbGciOiJIUzI1NiIsImV4cCI6MTQ2MDU5OTUxNCwiaWF0IjoxNDYwNTk1OTE0fQ.eyJpZCI6Mn0.FSoeJkqaV1Zlc1XjDu5fcI3fmRSHD1OMhm-M8sKOHE8:x
      
      parameters:
        - name: payload
          in: body
          description: Input Payload
          required: true
          schema:
            $ref: '#/definitions/AnswerPutInputPayload'
      tags:
        - Answer
      responses:
        200:
          description: Success Message
          schema:
            $ref: '#/definitions/AnswerPutOutputPayload'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'

##############################################
# Definitions
##############################################

definitions:

##############################################

  UserPostInputPayload:
    properties:
      username:
        type: string
        description: Username of a user
      password:
        type: string
        description: Password of a user

  UserPostOutputPayload:
    properties:
      Location:
        type: string
        description: Unique url endpoint for a user
      Username:
        type: string
        description: Username that was registered

##############################################

  UserGetInputPayload:
    properties:
      duration:
        type: integer
        format: int32
        description: Access token validation in seconds. Default is 10 minutes

  UserGetOutputPayload:
    properties:
      User:
        type: array
        items:
          $ref: '#/definitions/User'
      Token:
        type: array
        items:
          $ref: '#/definitions/Token'
          
  User:
    properties:
      name:
        type: string
        description: Name of the user
      rating:
        type: integer
        format: int32
        description: User rating from 0-5.
      tags:
        type: array
        description: Array of Tags associated with the user
        items:
          type: string
          description: Tag name. Currently supporteda are, "autry", "dbpedia","npg","saam","ulan","viaf"
      username:
        type: string
        description: Username of the user
  
  Token:
    properties:
      duration:
        type: integer
        format: int32
        description: Token validity in seconds
      value:
        type: string
        description: Alphanumeric value of a token

##############################################

  UserPutInputPayload:
    properties:
      name:
        type: string
        description: Name of the user
      rating:
        type: integer
        format: int32
        description: Rating of the user
      tags:
        type: array
        description: Array of Tags associated with the user
        items:
          type: string
          description: Tag name. Currently supporteda are, "autry", "dbpedia","npg","saam","ulan","viaf"
  
  UserPutOutputPayload:
    properties:
      name:
        type: string
        description: Name of the user
      rating:
        type: integer
        format: int32
        description: User rating from 0-5.
      tags:
        type: array
        description: Array of Tags associated with the user
        items:
          type: string
          description: Tag name. Currently supporteda are, "autry", "dbpedia","npg","saam","ulan","viaf"
      username:
        type: string
        description: Username of the user

##############################################        
  
  QuestionGetInputPayload:
    properties:
      stats:
        type: boolean
        description: Flag to request question stastics in the response
      bulk:
        type: integer
        format: int32
        description: Limit number of questions returned. Default is 10
        
  QuestionGetOutputPayload:
    properties:
      Exact_Match:
        type: array
        items:
          $ref: '#/definitions/Exactmatched'
      Unmatched:
        type: array
        items:
          $ref: '#/definitions/Unmatched'
      qid:
        type: string
        description: Unique question ID
      stats:
        type: string
        description: Returns question statitics if requested in query
        
  Exactmatched:
    properties:
      name:
        type: string
        description: Name of the property that matched exactly
      value:
        type: string
        description: Value of the property that matched exactly on both side.
        
  Unmatched:
    properties:
      name:
        type: string
        description: Name of the property.
      lvalue:
        type: string
        description: Value of the property on left side.
      rvalue:
        type: string
        description: Value of the property on right side.
        
##############################################        

  QuestionPostInputPayload:
    properties:
      uri1:
        type: string
        description: One of the URI of a pair of URI. Ordering does not matter. Used only for non-bulk reqeust.
      uri2:
        type: string
        description: The other URI of a pair of URI. Ordering does not matter. Used only for non-bulk reqeust.
      dedeupe:
        type: array
        description: List of annotations generated from dedupe. Used only for non-bulk reqeust.
        items:
          $ref: '#/definitions/dedupe'
      bulk:
        type: integer
        format: int32
        description: Indicates a bulk request and number of questions sent in a request. Used only for a bulk request and payload array of size = bulk is expected as well.
      payload:
        type: array
        description: List of annotations generated from dedupe. Used only for non-bulk reqeust.
        items:
          $ref: '#/definitions/questionPayload'
  
  dedupe:
    properties:
      fields:
        type: array
        description: List of fields used by dedupe
        items:
          type: string
          description: Field name. Few examples are, "schema:name", "schema:birthDate", "schema:deathDate"
      version:
        type: string
        description: Dedupe version that generated this annotation
      linkScore:
        type: string
        description: linkScore computed by dedupe for this question
  
  questionPayload:
    properties:
      uri1:
        type: string
        description: One of the URI of a pair of URI. Ordering does not matter. Used only for non-bulk reqeust.
      uri2:
        type: string
        description: The other URI of a pair of URI. Ordering does not matter. Used only for non-bulk reqeust.
      dedeupe:
        type: array
        description: List of annotations generated from dedupe. Used only for non-bulk reqeust.
        items:
          $ref: '#/definitions/dedupe'
          
  QuestionPostOutputPayload:
    properties:
      curation:
        type: array
        items:
          $ref: '#/definitions/curation'
        
  curation:
    properties:
      comment:
        type: string
        description: comment added with this decision value
      value:
        type: integer
        format: int32
        description: Decision value from 1 to 3
  
  ##############################################
  
  AnswerPutInputPayload:
    properties:
      qid:
        type: string
        description: qid of a question being answered
      comment:
        type: string
        description: Comment related to decision
      value:
        type: integer
        format: int32
        description: Decision value from 1 to 3
        
  AnswerPutOutputPayload:
    properties:
      message:
        type: string
        description: Message indicating that answer was submitted successfully
  
  ##############################################
  
  Error:
    properties:
      message:
        type: string
        description: Message related to error.
        